requirements-test.txt
# test time dependencies for jupyter_lsp, including one language server
# nodejs >=8
-r requirements.txt
flake8 >=3.5
pytest-asyncio
pytest-cov
pytest-flake8
pytest-runner
python-language-server
